+++ libtvvideoextractor/src/main/java/com/tajmoti/libtvvideoextractor/module/StringUtils.kt
@@ -1,0 +1,7 @@
+package com.tajmoti.libtvvideoextractor.module
+
+fun String.substringBetween(startSequence: String, endSequence: String): String {
+    val start = indexOf(startSequence) + startSequence.length
+    val end = indexOf(endSequence, start + 1)
+    return substring(start, end)
+}